refactor(syscall tests): Split syscall tests to be less clutered (#1961)
