+++ packages/neo-one-node-vm/src/__data__/index.ts
@@ -5,5 +5,6 @@ import * as testUtils from './utils';
 export * from './createBlockchain';
 export * from './keys';
 export * from './factory';
+export * from './runSysCalls';
 
 export { assets, testUtils, transactions };
